requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
dataclasses-json==0.6.7
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.2
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.15.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
keyring==24.3.1
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.8
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.79
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.34.0
orjson==3.10.5
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlmodel==0.0.19
starlette==0.37.2
tenacity==8.4.1
tiktoken==0.7.0
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
trove-classifiers==2024.5.22
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
xattr==1.1.0
yarl==1.9.4
zipp==3.19.2
